requirements.txt
DX.py==0.3.1
dxpy==0.320.0
Flask==2.2.3
matplotlib==3.7.1
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.2
pytz==2022.6
PyYAML==5.3.1